More segmentation and clean up.
Starting to rework the configuration a little more to match mcrouter,
which is the ultimate inspiration of writing a replacement for
twemproxy, to be honest.

Starting to split up the code to match a more natural flow of.... for
each config, give us back a future that represents the entire pool...
and each pool will be based on its own futures representing sockets and
backends and so forth.