Big performance win!  Within striking distance of Twemproxy now.
This is all due in part to batched up received messages/outgoing
messages to be able to work with pipelining.  It also looks like the
copypasta job where I stole some existing code to formulate the batch
combinator was giving subpar performance.  By ensuring that our internal
buffer for the combinator was getting replaced with a properly-sized --
matching the existing size instead of from 0 -- new buffer, performance
shot up by a lot.

I still need to do some careful examination to make sure we're not
needlessly allocating, and I'm also not sure if there's anywhere we
could stand to use buffer/object pools, but I think we can get back to
actual feature parity work for the time being.

Hooray!