Reorganization to try and support routing strategies.
Right now it's ugly because it's a lot of boxing and nonsense, but, I'm
trying to start to lay the groundwork for different routing strategies
like... a "normal" mode where requests just go to one backend pool and
responses get shipped back vs a "warm up" mode where we send all
requests to a secondary back up pool just to warm it up but discard
everything it sends back, etc.

For Redis, a lot of things that, say, mcrouter has for routes just don't
necessarily work, but I think the biggest things we care about are
traffic shadowing/warm up, so that alone over Twemproxy would already be
very nice to have.

We also had to reorganize a bit to ship errors during listener
configuration back to main and to make sure logging properly flushes in
those cases, as well.