Basic support for recovering from a failed backend connection.
Basic support is a weird wy to put it, because it works totally fine...
but functionally, the way we recover doesn't match the behaviour of
Twemproxy insofar as we still send back the error -- failed write/read
to backend, yadda -- to the client where Twemproxy seems to retry the
request by opening a new backend connection or w/e.

Practically speaking, this means redis-benchmark can run uninterrupted,
or that clients would just see elevated response timings during network
blips.  I'm _pretty_ sure retrying the backend would be the right choice,
but I just need to find the right way to do it.