Add support for properly returning errors from backends to clients.
This will, on the batched write-level, return errors to the client.
When the batches are actual batches, and not just a single operation,
all operations in the batch for a given backend will be failed if any
failure occurs.  This is technically suboptimal but I think a reasonable
behavior.

This _won't_ cause all realized batches to fail, though.  For example,
if we run 6 operations in total, over 3 batches, and one batch fails,
those two "slots" in the pipelined responses will both be errors, even
if only one failed, but the other four slots for the other two batches
will be whatever the server responded back with.