Big fix to make our backend connection queue not stall clients.
The previous design was not resignalling tasks who we had told that no
connections were available aka Async::NotReady.  Due to the way the
runtime works, these tasks would never again poll the backend stream
unless notified, so one connection would essentially be returning the
connect to itself but it never got the not ready signal, and chew
through very fast with no lock contention.

Unfortunately, this means we now are now contending heavily over the
backend synchronization, and even just the overhead of the machinery in
place to do so has reduced our throughput drastically.  The more
clients, the worse it gets, whereas Twemproxy's performance gets better
with more clients, to a point.

We'll have to see what can be done here.