Start of a barebones integration test suite for Synchrotron.
In the spirit of Twemproxy, I'm building a standalone test suite that
can run integration-level tests against Synchrotron.

It controls both Synchrotron itself and the underlying cache servers,
starting and stopping discrete instances and configuring them as needed
for the tests.

This will give us a relatively easy way to test things end-to-end in a
controlled fashion.