Switch to futures-turnstyle and prepare for stats.
We're using futures-turnstyle now instead of rs_futures_spmc to signal
to our listeners to close.  This will pave the way for us to do online
reconfiguration with minimal signaling code.

Also starting to think about what stats will look like.  Futures, as of
now, are dominated by combinators that assume ownership, shared-nothing,
and metrics are not easy to weave into that, especially with logic
spanning multiple combinators that encompasses an operation to be timed.