First pass at adding in metrics support.
We've only instrumented the "normal" Redis handler, but I think it's
looking good so far.  The interface isn't the greatest, but it's
workable.  It shouldn't be hampering performance, not nearly as much as
the design itself would be accounting for, so that's good.

Time to start instrumenting more points so we can optimize!