Introduction of a routing layer.
One of the major goals of Synchrotron was to take the best of both
mcrouter and Twemproxy and brind them together.  In mcrouter, the
premiere feature is the routing capabilties.  This is the ability to
define how requests are handled from clients, and mcrouter's
capabilities are very flexible.

Now we have this for Synchrotron! Or, at least, the groundwork for it.
The initial cut of this is a "fixed" router: all the requests that come
into it are handed off to backends, responses handed back.

The expansion of this is pretty straight forward: routers can (and
will!) compose themselves to build flexible routing chains, and this
will allow operators to execute on both simple and complex topologies.