Stage one of both short-circuit messages and graceful closing.
We've introduced some new Redis "message types" that correspond directly
to handling pings and handling quits.

Handling pings is a short circuit: we don't want to actual query a
backend server.  In practice, this doesn't short circuit until it gets
to the point of trying to send to a backend, but nothing will actually
be sent.  Even though we do a 0-byte send to a backend, it's still much
faster in the uncontended case.

The biggest problem with this design is that we essentially still tie up
the backend connection running the future which ultimately attempts the
0-byte read.  On top of that, because we're assigning the message to a
given backend based on hashing the message, the hash "key" will always
be "ping", so one backend in particular would be the destination for all
pings.  I'd be surprised if somehow a caching cluster was taking
significant ping traffic, but theoretically speaking, this is
suboptimal.

The hard part is that our abstract design forces us into
tying up a backend connection very early in the request handling, so
getting it to shortcircuit before that happens could be tricky.  There
might be a chance that we could roll in the batch combinator to the
message stream directly, and only ever return a batch of 1 when we
detect a ping, sort of like a line break... and we'd just handle that
batch outside of the routing machinery.  Not sure yet.

Quits are a little different, but simpler.  Stage two of this work will
be to make the message stream terminate when it detects a quit message,
so that we can stop parsing incoming data, but still get all of the
in-flight requests (including quit, since technically you get back an
+OK from the server) processed and sent back to the client.  This is
simple to do, I just haven't done it yet.

We're also starting to bake more integration tests which will be nice to
start to prove out the end-to-end style behavior.