Added support to close a message stream when we receive a quit message.
This does work, although the error message we get back from the redis
crate is weird: not enougb bytes to read.  The error isn't wrong, but I
was hoping for/expecting some sort of "connection dropped" or
"connection reset by peer" type of thing.  I guess either way, it
acheives the intended effect. ¯\_(ツ)_/¯