Rework how we handle processing requests and responses.
This represents a change to our internal forwarding architecture, bring
it more inline with how Twemproxy works.

Before this, the request/response flow was heavily serialized.  A
client's request would come in, be potentially batched, and then be
batched at a per-backend level, and would run against the backends...
waiting for all responses to come back before finally returning the
responses to the client.

This worked well enough for low client counts when contention for the
backend was low, but ultimately is a scaling bottleneck.

As mentioned above, we've changed our architecture to reflect Twemproxy,
which uses a sort of... message queue-meets-reserved-slots-in-an-array
approach.

When requests are read from a client, before we actually run them, we
reserve "slots" in the client's message queue.  These slots in
themselves are faux futures, representing the space for the response to
a given request. (which also codifies its order for us implicitly)

Requests are wrapped to include this message queue information, so that
when they finally make it to a backend and the response is read from the
server, we simply pass it back to the message queue, which handles
defragmenting and ordering for us.  This does two major things for us:

First, our routing layer no longer has to care about message
fragmentation.  It was already black boxed by the request processors
themselves, but it was a large chunk of code in the router that just
didn't matter.  Now that we've moved it to the message queue, the router
is simpler and we have better chances to start sending data to the
client as it arrives, not only after we've gotten it all.

Secondly, this also allows us to push requests into the backends more
freely.  While not implemented -- although it will only require a few
lines of code -- this work has unlocked the ability for us to run
requests for more than a single client at a time per backend connection.
Previously, backend connections were only ever able to be used by a
single client, due to the nature of how we bound them in the large
future chain that drove every client request batch.  This represents a
huge concurrency opportunity for us.

There's still a lot of work, IMO, to clean this up and simplify it
further.  We're doing cheap-and-easy things right now: current thread
runtime, spawning everything separately instead of intelligently
combining futures to reduce context switching, etc.  Overall, though,
we're on more scalable groundwork than we were before.