Break down the test code into better chunks.
We're also using a static atomic usize now to handle tracking the port
offset, so we no longer need to pass it in the tests when we want our
daemons, which is wayyy better.  We also provide the connection string
as a method on the handles, to make it more contained.