Add inline message support. (#7)
* Make stats port configurable.

This was in the configuration schema before, but we just weren't using
it.  It's important because without a different port, it messes up our
multithreaded integration tests.

* Add inline message support.

This new message state allows us to enqueue messages which themselves as
the actual response.  This is important for things like pings because we
really don't want to have to go all the way through the backend
processing chain.