Add the ability to reload listeners at runtime.
We listen for the USR1 signal, which when fired, will trigger a new
'supervisor' future to reread the configuration and spawn all of the
configured listeners.  When those listeners have been spawned, the old
listeners will be notified to close.

Problem(s) to fix:
Currently, you can ctrl-c and have the server quickly close, even with
an active connection.  This shouldn't happen; listeners should stop but
existing connections should be able to drain.  Obviously we don't want
intentional/unintentional Slowloris-y connections to mess us up and
prevent closing, but we definitely want to let things finish.  We still
need to look into this.

Beyond that, I need to verify that the reload behavior doesn't have a
window of connections not being accepted.  Per the documentation,
binding a `TcpListener`, and having it return `Ok(_)` is supposed to
indicate the socket was successfully bound, and should mean it's
capturing accepts.  We bind before we spawn, and we spawn before we
close the old listeners, so this _should_ mean that we're never missing
new connections... but I still want to write a test application or
something to stress test that invariant.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>