Switch to using tokio-evacuate to provide our tiered close signal.
Originally, we were using a single shared turnstyle waiter to close
everything.  The problem with this approach is that it had us closing
all of the necessary bits -- backend pools, backends themselves, etc --
that you would need to actually service a request.  So while we wanted
to let existing clients, who connected before the listener closed, a
chance to finish, the necessary bits went away and closed all of the
connections.

Now, we use tokio-evacuate to provide a tiered close: the "global" close
will close listeners to prevent new connections, and tokio-evacuate wil
trigger off of that to give clients a window of time to finish their
work or be forcefully closed.  All components such a the backend pool,
backends, etc, work off of the "secondary" close signal.