Add support for traffic shadowing. (#33)
* core: make QueuedMessage clonable in read-only mode

In preparation for taking a single request and splitting it, we've
changed QueuedMessage so that it can be cloned via `as_read`, which
returns a read-only copy.  It copies the internal message, but
responding to this message via `respond` is essentially a no-op.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* protocol: fixups to redis to prepare for shadowing

We needed some machinery to send commands to the server to prevent
server replies, and `from_inline` only worked for non-RESP commands, so
we made it more legit.

Also created a helper future type to wrap futures that deal with
utilizing a backend connection.  This should help us eliminate a lot of
type constraints on the processor.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* processor: add support for "noreply" mode in processor

We've now baked in support for "noreply" mode in processors.  This
allows backends to choose whether or not they want to run a request and
get back a response or not.  This is a core part of traffic shadowing in
that we want to send all requests to a shadow backend but don't actually
_need_ the responses.

We've also switched to using ProcessFuture and have dropped a _lot_ type
constraints on the processor - woohoo!

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* tests: add new test for traffic shadowing

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* core: fix up noreply and rename RequestProcessor to Processor.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>